+++ z/util/regex/rules/punct.hpp
@@ -24,7 +24,7 @@ namespace z
 				 * \param max The maximum number of times this rule can match.
 				 * \param greedy Whether this rule should consume input greedily.
 				 */
-				punct(bool negate=false, size_t min=1, size_t max=1, bool greedy=true);
+				punct(bool negate=false, int min=1, int max=1, bool greedy=true);
 				bool match(uint32_t current) const;
 
 				///If true, only matches non-punctuation characters.
