+++ z/util/regex/rules/range.hpp
@@ -22,7 +22,7 @@ namespace z
 				 * \param max The maximum number of times this rule can match.
 				 * \param greedy Whether this rule should consume input greedily.
 				 */
-				range(uint32_t begin, uint32_t end, bool insensitive=false, size_t min=1, size_t max=1, bool greedy=true);
+				range(uint32_t begin, uint32_t end, bool insensitive=false, int min=1, int max=1, bool greedy=true);
 
 				bool match(uint32_t current) const;
